requirements.txt
# Core ML/DL
transformers
sentence-transformers
faiss-cpu
faster-whisper
llama-cpp-python
huggingface_hub

# Application Frameworks
gradio
Flask
requests

# File Processing
Pillow
pypdf
librosa
soundfile
python-docx
pandas

# Utilities
numpy
